downloader: improve UlrSave, Check if the file exists before downloading

+++ extractors/utils/utils.go
@@ -4,6 +4,6 @@ type VideoData struct {
 	Site  string
 	Title string
 	Url   string
-	Size  int
+	Size  int64
 	Ext   string
 }
